Fix bbox bug

+++ lib/src/Backend/D3D9/AtomsphereRendererDX9.h
@@ -13,7 +13,7 @@ private:
     String ShaderHeader(const AtmosphereParams &params) const;
     String TransmittanceShader() const;
 
-    float Interpolate(const CArray<float> &arrWaveLength, const CArray<float> &arrWaveLengthFunction, float WaveLength) const;
+    double Interpolate(const CArray<double> &arrWaveLength, const CArray<double> &arrWaveLengthFunction, double WaveLength) const;
 
 private:
     IDirect3DTexture9 *m_pTransmittance = nullptr;
